fix: Correct Player package import and API method calls
- Change Player import to com.hypixel.hytale.server.core.entity.entities.Player
- Replace getUsername() with getName() for Player
- Fix PlayerDisconnectEvent player access
- Fix double to int conversion in SkillManager
- Remove AdminCommand temporarily (missing API methods)

+++ src/main/java/de/nightraid/nrcrpg/managers/SkillManager.java
@@ -30,7 +30,7 @@ public class SkillManager {
         int currentLevel = skillData.getLevel();
         double currentXP = skillData.getXp();
         
-        int requiredXP = XPCalculator.getRequiredXP(currentLevel);
+        double requiredXP = XPCalculator.getRequiredXP(currentLevel);
         
         if (currentXP >= requiredXP && currentLevel < 100) {
             skillData.levelUp();
